CoinValuator can make change

+++ vending-machine/src/main/groovy/com/mholtman/domain/Product.groovy
@@ -4,6 +4,6 @@ import groovy.transform.EqualsAndHashCode
 
 @EqualsAndHashCode
 class Product {
-    double price
+    Integer price
     String name
 }
